Install accents-substitute module
That's a conveninece module which exports accents-substitute-latin1
and accents-substitute-utf8 from their respective modules.

As a intended side-effect, this change fixes a problem that would make
chicken-install always resintall this egg when "accents-substitute" is
in the dependencies list of some other egg.  Since accents-substitute
didn't install an accents-substitute module (it only installed
accents-substitute-latin1 and accents-substitute-utf8),
chicken-install would not detect it is installed.  On the other hand,
you can't use accents-substitute-latin1 or accents-substitute-utf8 in
dependencies lists because they are not eggs (they are modules of
eggs).